ToDo Recycler on Room фильтр обновляется Add Delete итд  работает Clear create in menuBar количество записей в заголовке Repository ToDo03 Card + Menu DataConverters Canary 8 Date Int + Char Index MemoryBase Old+Pager_Cheese Pager Ready ToDo SQL PAEMI - Int Id Int Calendar

+++ ToDo_SQL/app/src/main/java/com/app4web/asdzendo/todo/database/FactDatabaseDao.kt
@@ -65,7 +65,7 @@ interface FactDatabaseDao {
      * @param id startTimeMilli to match
      */
     @Query("SELECT * from fact_todo WHERE factId = :id")
-    fun get(id: Long): Fact?
+    fun get(id: Int): Fact?
 
     /**
      * Deletes all values from the table.
@@ -113,7 +113,7 @@ interface FactDatabaseDao {
      * Выбирает и возвращает fact c id.
      */
     @Query("SELECT * from fact_todo WHERE factId = :id")
-    fun getFactWithId(id: Long): LiveData<Fact>
+    fun getFactWithId(id: Int): LiveData<Fact>
 
     /**
      * выборка фильтром PAEMI в обратной сортировке
