+++ ToDo_SQL/app/src/main/java/com/app4web/asdzendo/todo/ui/detail/FactDetailViewModel.kt
@@ -9,7 +9,7 @@ import timber.log.Timber
 
 class FactDetailViewModel(
         private val factRepository: FactRepository,
-        factID: Long = 0L,
+        factID: Int = 0,
         paemi: Int = 0,
 ): ViewModel() {
     init { Timber.i("TODO FactDetailViewModel created $factID")}
@@ -34,7 +34,7 @@ class FactDetailViewModel(
     //var paemistring = getFact().value.paemi //PAEMI.values()[(getFact().value!!).paemi].name  // Временно для TextView поменять на адаптер
 
     init {
-        val fact0L: LiveData<Fact> = if (factID == 0L)
+        val fact0L: LiveData<Fact> = if (factID == 0)
                 MutableLiveData(Fact(paemi = paemi, nameShort = "новый Факт", name = "Факт полностью: новый"))
                  else factRepository.getFactWithId(factID)
         fact.addSource(fact0L, fact::setValue)
